refactor(tests): organize interpreter array tests
